Clean up + minor refactors including improving perms loading state in storage settings
