--- apps/studio/components/interfaces/Storage/index.ts
@@ -1,5 +1,0 @@
-export { default as StorageExplorer } from './StorageExplorer/StorageExplorer'
-export { default as StoragePolicies } from './StoragePolicies/StoragePolicies'
-export { default as StorageSettings } from './StorageSettings/StorageSettings'
-
-export { DeleteBucketModal } from './DeleteBucketModal'
